+++ tests/test_EightDotThree.py
@@ -129,10 +129,11 @@ def test_make_8dot3_name_unicode():
     """Test that make_8dot3_filename generates valid 8dot3 filenames."""
     fde = mock.MagicMock()
     fde.get_entries.return_value = ([], [], [])
-    sfn = EightDotThree()
-    lfn = sfn.make_8dot3_name("🤷.🤷", fde)
-    assert "_._" == lfn
-    assert sfn.is_8dot3_conform(lfn)
+    sfn = EightDotThree(encoding='UTF-8')
+    strname = sfn.make_8dot3_name("🤷.🤷", fde)
+    assert "🤷.🤷" == strname
+    assert not sfn.is_8dot3_conform(strname)
+    assert sfn.is_8dot3_conform(strname, encoding='UTF-8')
 
 
 def test_make_8dot3_name_collision():